fix: rename screen files
