+++ lib/routes/app_router.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:wesell/views/balance/balance_screen.dart';
-import 'package:wesell/views/identity/verification_method.dart';
+import 'package:wesell/views/identity/verification_method_screen.dart';
 import 'package:wesell/views/pin/new_pin_1_screen.dart';
 import 'package:wesell/views/pin/new_pin_2_screen.dart';
 import 'package:wesell/views/pin/old_pin_screen.dart';
@@ -10,7 +10,7 @@ import '../views/auth/login_screen.dart';
 import '../views/home/home_screen.dart';
 import '../views/item/item_screen.dart';
 import '../views/chat/chat_screen.dart';
-import '../views/me/edit_username.dart';
+import '../views/me/edit_username_screen.dart';
 import '../views/me/edit_avatar_screen.dart';
 import '../views/me/me_screen.dart';
 import '../providers/auth_provider.dart';
